+++ freepaths/Legacy/lattices.py
